Move the `getExternalConfig()` function to a better home
It really needs to be reusable instead of being stuck in the script. In
particular since we're about to modify GitGitGadget such that it can be
run as a GitHub Action instead.
