+++ lib/project-config.ts
@@ -64,6 +64,25 @@ export function getConfig(): IConfig {
     return config;
 }
 
+export async function getExternalConfig(file: string): Promise<IConfig> {
+    const filePath = path.resolve(file);
+    const newConfig = await loadConfig(filePath);
+
+    if (!Object.prototype.hasOwnProperty.call(newConfig, "project")) {
+        throw new Error(`User configurations must have a 'project:'.  Not found in ${filePath}`);
+    }
+
+    if (!newConfig.repo.owner.match(/^[a-z\d](?:[a-z\d]|-(?=[a-z\d])){0,38}$/i)) {
+        throw new Error(`Invalid 'owner' ${newConfig.repo.owner} in ${filePath}`);
+    }
+
+    if (!newConfig.repo.baseOwner.match(/^[a-z\d](?:[a-z\d]|-(?=[a-z\d])){0,38}$/i)) {
+        throw new Error(`Invalid 'baseOwner' ${newConfig.repo.baseOwner} in ${filePath}`);
+    }
+
+    return newConfig;
+}
+
 type importedConfig = { default: IConfig };
 
 /**
